Fixed types for vertices in tests for sin and cos #597 #598

+++ eo-runtime/src/test/java/EOorg/EOeolang/EOfloatEOcosTest.java
@@ -17,7 +17,7 @@ public final class EOfloatEOcosTest {
         MatcherAssert.assertThat(
             new Dataized(
                 new EOfloat$EOcos(
-                    new Data.ToPhi(0)
+                    new Data.ToPhi((double) 0)
                 )
             ).take(Double.class),
             Matchers.equalTo(cos0)
